requirements.txt
einops==0.8.0
imageio==2.33.1
numpy==2.1.1
Pillow==10.4.0
tensorboardX==2.6.2.2
torch==2.2.1
torch_scatter==2.1.2+pt22cu121
torchvision==0.17.1
